fix(quota): keep fractional usagePercent decimals (opencode upstream now sends 19.3)
opencode.ai server-fn:3 switched usagePercent from integers to decimals
(19.3/18.5/71.3 + usage/limit byte totals). parseNonNegInt (digits-only)
failed on '.', erroring the whole account with 'invalid usagePercent'.

- QuotaUsage.UsagePercent int -> float64 (shared type; commandcode/
  ollama drop their math.Round and preserve precision too; kimi already
  documented this decision for 2.19/10.42)
- opencode parseUsageObject uses parseNonNegPercent, keeps exact value
- frontend percentPrecision 0->1 for opencode/commandcode/ollama rows
- tests: fractional accept + malformed reject fixtures; commandcode/
  ollama assertions updated to exact decimals
- version bump 0.11.0 -> 0.11.1

+++ internal/formatter/format.go
@@ -27,14 +27,18 @@ func FormatPercent(value int) string {
 	return fmt.Sprintf("%d%%", value)
 }
 
-func ProgressBar(percent int, width int) string {
-	filled := (percent * width) / 100
+// UsagePercent is float64 (not int) so fractional percents from upstream
+// payloads survive rendering — same decision KimiQuotaUsage documents for
+// 2.19/10.42. The bar's filled slots round integer; the trailing label
+// keeps one decimal.
+func ProgressBar(percent float64, width int) string {
+	filled := int(percent*float64(width)) / 100
 	var sb strings.Builder
 	sb.Grow(width + 6)
 	sb.WriteByte('[')
 	for i := 0; i < width; i++ {
 		if i < filled { sb.WriteString("█") } else { sb.WriteString("░") }
 	}
-	sb.WriteString(fmt.Sprintf("] %d%%", percent))
+	sb.WriteString(fmt.Sprintf("] %.1f%%", percent))
 	return sb.String()
 }
